+++ src/utils/transformers/check-in.ts
@@ -1,15 +1,15 @@
 import { EventResponse } from '@/types/dataModel/event';
 import { CreateEventVolunteerRequest } from '@/types/dataModel/eventVolunteer';
-import { VolunteerResponse } from '@/types/dataModel/volunteer';
+import { CreateVolunteerRequest } from '@/types/dataModel/volunteer';
 import { CheckInFormData } from '@/types/forms/checkIn';
 
 export function transformCheckInFormDataToCreateEventVolunteerRequest(
   formData: CheckInFormData,
-  volunteer: VolunteerResponse,
+  volunteer: string | CreateVolunteerRequest,
   event: EventResponse
 ): CreateEventVolunteerRequest {
   return {
-    volunteer: volunteer._id,
+    volunteer,
     event: event._id,
     role: formData.role,
     organization: formData.organization?._id,
